[UPDATE] update InPlan API, add Mock Test

+++ src/main/java/com/inplan/inplan/InPlanService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface InPlanService {
     void putUser(User user);
-    List<User> getUserById(int id);
-    void deleteUserById(int id);
-    void updateUserById(int id, User user);
+    List<User> getUserById(Long id);
+    void deleteUserById(Long id);
+    void updateUserById(Long id, User user);
 
 }
