+++ src/main/java/com/inplan/inplan/repository/PlanCategoryRepository.java
@@ -3,5 +3,5 @@ package com.inplan.inplan.repository;
 import com.inplan.inplan.dao.PlanCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanCategoryRepository extends JpaRepository<PlanCategory, Integer> {
+public interface PlanCategoryRepository extends JpaRepository<PlanCategory, Long> {
 }
